JIT native interface.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1187 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcretest.c
@@ -4415,7 +4415,7 @@ while (!done)
       dbuffer = (pcre_uint8 *)realloc(dbuffer, dbuffer_size * CHAR_SIZE);
       if (dbuffer == NULL)
         {
-        fprintf(stderr, "pcretest: malloc(%d) failed\n", dbuffer_size);
+        fprintf(stderr, "pcretest: malloc(%d) failed\n", (int)dbuffer_size);
         exit(1);
         }
       }
